Convert logger.js to TypeScript and enhance debug print
