+++ deploy/index.ts
@@ -5,8 +5,7 @@ import {PassThrough} from 'stream';
 import concat from 'concat-stream';
 import {FastifyInstance} from 'fastify';
 import {get} from 'lodash';
-// @ts-ignore
-import logger from '../lib/logger.js';
+import logger from '../lib/logger';
 import type {SlackInterface} from '../lib/slack';
 
 // @ts-ignore
