+++ oauth/index.ts
@@ -1,7 +1,6 @@
 import {FastifyInstance} from 'fastify';
 import type {SlackInterface, SlackOauthEndpoint} from '../lib/slack';
-// @ts-ignore
-import logger from '../lib/logger.js';
+import logger from '../lib/logger';
 import * as sqlite from 'sqlite';
 import sqlite3 from 'sqlite3';
 import path from 'path';
